requirements.txt
Pillow>=10.0.0
requests>=2.31.0
cairosvg>=2.7.0
boto3>=1.28.0
psycopg2-binary>=2.9.0
auth0-python>=4.0.0 